finsihed with code optimization :tada: :tada: :smiley:
