+++ contact.html
@@ -20,10 +20,10 @@
     <link rel="stylesheet" href="./styles/styles.css" />
   </head>
   <body>
-    <header class="contact-header">
+    <header>
       <div class="top-bar">
         <div class="wrapper">
-          <h3>Hello Everyone! 25% off all products</h3>
+          <h3>Hello World! 25% off all products</h3>
           <ul>
             <li><a href="#">Cad</a></li>
             <li><a href="#">English</a></li>
@@ -33,7 +33,7 @@
       <nav class="navbar">
         <div class="wrapper">
           <div class="logo">
-            <img src="./assets/logo.png" alt="logo pic" />
+            <img src="./assets/logo.png" alt="logo" />
           </div>
 
           <ul class="nav" id="nav">
@@ -52,13 +52,20 @@
               src="./assets/icons/shopping-bag.svg"
               width="24"
               height="24"
-              alt="icon of shopping bag"
+              alt="shopping-bag icon"
             />
-            <span>3</span>
+            <span id="total-items">0</span>
           </div>
           <button type="button" class="hamburger">&#9776;</button>
         </div>
       </nav>
+
+      <div class="cart-design-container">
+        <h2 class="cart-heading"></h2>
+        <ul class="cart-info"></ul>
+        <div class="checkout-price-info"></div>
+        <div class="cart-buttons"></div>
+      </div>
 
       <div class="contact-header-banner">
         <div class="wrapper">
@@ -240,6 +247,5 @@
       </div>
     </footer>
   </body>
-  <script src="./scripts/toggleMenu.js"></script>
-  <script src="./scripts/formValidations.js"></script>
+  <script src="./scripts/contact.js" type="module"></script>
 </html>
