--- scripts/toggleMenu.js
@@ -1,8 +1,0 @@
-// Toggle Menu
-const hamburger = document.querySelector(".hamburger");
-const navMenuBar = document.querySelector(".nav");
-const shoppingBag = document.querySelector(".shopping-bag");
-const toggleMenu = hamburger.addEventListener("click", () => {
-  navMenuBar.classList.toggle("nav-sm");
-  shoppingBag.classList.toggle("nav-sm");
-});
